test type 5

+++ api/controllers/AssessmentsController.php
@@ -121,8 +121,7 @@ class AssessmentsController extends  MyActiveController
 
         if(!$survey_done)  return ResponseHelper::sendFailedResponse(['message'=>'Survey is Completed'],404);
         
-        if (!isset($params['answers']) || !is_array($params['answers'])){
-            return isset($params['answers']);
+        if (!array_key_exists('answers',$params) || !is_array($params['answers'])){
             return ResponseHelper::sendFailedResponse(['message'=>'Invalid Params'],400);
         };
 
@@ -189,54 +188,53 @@ class AssessmentsController extends  MyActiveController
                 $userAnswer->survey_user_answer_point = $point;
                 $userAnswer->survey_user_answer_value = $value;
                 $userAnswer->save(false);
-              }elseif ($question->survey_question_type === SurveyType::TYPE_DROPDOWN){
-                 $answerPoint = SurveyAnswer::findOne(['survey_answer_id'=>$value]);
-                 $userAnswers = $question->userAnswers;
-                 $userAnswer = !empty(current($userAnswers)) ? current($userAnswers) : (new SurveyUserAnswer([
-                     'survey_user_answer_user_id' => \Yii::$app->user->getId(),
-                     'survey_user_answer_survey_id' => $question->survey_question_survey_id,
-                     'survey_user_answer_question_id' => $question->survey_question_id,
-                 ]));
-                if ($answerPoint->correct) {
-                    $userAnswer->survey_user_answer_point = $question->survey_question_point;
-                 }else{
-                    if ($params['status'] == 2) {
-                        $this->correctiveActionReport($question,$answerPoint);
-                    }
-                 }
-                $userAnswer->survey_user_answer_answer_id = $value;
-                $userAnswer->survey_user_answer_value = $value;
-                $userAnswer->save(false);
-
-              }elseif ($question->survey_question_type === SurveyType::TYPE_ONE_OF_LIST
-              ){
-                 //handel one answer
-                if (!is_array($value) and is_integer( (int) $value )) {
-                    // return $value;
-                    $answerPoint = SurveyAnswer::findOne(['survey_answer_id'=>$value]);
-                     if ($answerPoint) {
-                         $userAnswers = $question->userAnswers;
-                         $userAnswer = !empty(current($userAnswers)) ? current($userAnswers) : (new SurveyUserAnswer([
-                             'survey_user_answer_user_id' => \Yii::$app->user->getId(),
-                             'survey_user_answer_survey_id' => $question->survey_question_survey_id,
-                             'survey_user_answer_question_id' => $question->survey_question_id,
-                         ]));
-                         if ($answerPoint->correct) {
-                            $userAnswer->survey_user_answer_point = $answerPoint->question->survey_question_point;
-                         }else{
-                            if ($params['status'] == 2) {
-                                $this->correctiveActionReport($question,$answerPoint);
-                            }
-                         }
-                         $userAnswer->survey_user_answer_answer_id = $value;
-                         $userAnswer->survey_user_answer_value = $value;
-                         $userAnswer->save(false);
-                    }else{
-                        return ResponseHelper::sendFailedResponse(['message'=>'Invalid Params'],400);
-                    }   
+              }elseif ($question->survey_question_type === SurveyType::TYPE_DROPDOWN and !is_array($value) and is_integer( (int) $value )){
+                 $answerPoint = SurveyAnswer::findOne(['survey_answer_id'=>$value,'survey_answer_question_id'=>$question->survey_question_id]);
+                if ($answerPoint) {
+                     $userAnswers = $question->userAnswers;
+                     $userAnswer = !empty(current($userAnswers)) ? current($userAnswers) : (new SurveyUserAnswer([
+                         'survey_user_answer_user_id' => \Yii::$app->user->getId(),
+                         'survey_user_answer_survey_id' => $question->survey_question_survey_id,
+                         'survey_user_answer_question_id' => $question->survey_question_id,
+                     ]));
+                    if ($answerPoint->correct) {
+                        $userAnswer->survey_user_answer_point = $question->survey_question_point;
+                     }else{
+                        if ($params['status'] == 2) {
+                            $this->correctiveActionReport($question,$answerPoint);
+                        }
+                     }
+                    $userAnswer->survey_user_answer_answer_id = $value;
+                    $userAnswer->survey_user_answer_value = $value;
+                    $userAnswer->save(false);
                 }else{
                     return ResponseHelper::sendFailedResponse(['message'=>'Bad Request'],400);
                 }
+
+              }elseif ($question->survey_question_type === SurveyType::TYPE_ONE_OF_LIST
+               and !is_array($value) and is_integer( (int) $value )){
+                 //handel one answer
+                $answerPoint = SurveyAnswer::findOne(['survey_answer_id'=>$value,'survey_answer_question_id'=>$question->survey_question_id]);
+                 if ($answerPoint) {
+                     $userAnswers = $question->userAnswers;
+                     $userAnswer = !empty(current($userAnswers)) ? current($userAnswers) : (new SurveyUserAnswer([
+                         'survey_user_answer_user_id' => \Yii::$app->user->getId(),
+                         'survey_user_answer_survey_id' => $question->survey_question_survey_id,
+                         'survey_user_answer_question_id' => $question->survey_question_id,
+                     ]));
+                     if ($answerPoint->correct) {
+                        $userAnswer->survey_user_answer_point = $answerPoint->question->survey_question_point;
+                     }else{
+                        if ($params['status'] == 2) {
+                            $this->correctiveActionReport($question,$answerPoint);
+                        }
+                     }
+                     $userAnswer->survey_user_answer_answer_id = $value;
+                     $userAnswer->survey_user_answer_value = $value;
+                     $userAnswer->save(false);
+                }else{
+                    return ResponseHelper::sendFailedResponse(['message'=>'Bad Request'],400);
+                }   
               }else if($question->survey_question_type === SurveyType::TYPE_MULTIPLE
                  || $question->survey_question_type === SurveyType::TYPE_MULTIPLE_TEXTBOX
              ) {
@@ -253,10 +251,11 @@ class AssessmentsController extends  MyActiveController
                 }else{
                     $point = 0;
                 }
-                 // return var_dump($point);
+                 $valid_count = 0;
                  foreach ($question->answers as $i => $answer) {
                    $found = in_array($answer->survey_answer_id ,$value);
                     if($found){
+                        $valid_count++; 
                         $userAnswer =  new SurveyUserAnswer();
 
                             $userAnswer->survey_user_answer_user_id = \Yii::$app->user->getId();
@@ -273,24 +272,27 @@ class AssessmentsController extends  MyActiveController
                             }
                         $userAnswer->save(false);
                     }
+                }
 
-                   }
+                if (count($value) > 0 and $valid_count == 0) {
+                    return ResponseHelper::sendFailedResponse(['message'=>'Bad Request'],400);
+                }
 
-             }else if(
-                $question->survey_question_type === SurveyType::TYPE_RANKING
-             ) {
+             }else if($question->survey_question_type === SurveyType::TYPE_RANKING and is_array($value)) {
                  //delete old answers and add new
                  SurveyUserAnswer::deleteAll(['survey_user_answer_survey_id'=>$question->survey_question_survey_id ,
                      'survey_user_answer_question_id'=>$question->survey_question_id,
                      'survey_user_answer_user_id' => \Yii::$app->user->getId()
                      ]);
                  //save multiple
-                 foreach ($question->answers as $i => $answer) {
+                $valid_count = 0;
+                foreach ($question->answers as $i => $answer) {
                    $ids = array_keys($value);
                    $found = in_array($answer->survey_answer_id ,$ids);
                     if($found){
-                        $userAnswer =  new SurveyUserAnswer();
-
+                        if (isset($value[$answer->survey_answer_id]['rate'])) {
+                            $valid_count++; 
+                            $userAnswer =  new SurveyUserAnswer();
                             $userAnswer->survey_user_answer_user_id = \Yii::$app->user->getId();
                             $userAnswer->survey_user_answer_survey_id = $question->survey_question_survey_id;
                             $userAnswer->survey_user_answer_question_id = $question->survey_question_id;
@@ -299,12 +301,15 @@ class AssessmentsController extends  MyActiveController
                             if ($i == 0) {
                                 $userAnswer->survey_user_answer_point = $answer->question->survey_question_point;
                             }
-
-                        $userAnswer->save(false);
+                            $userAnswer->save(false);
+                        }else{
+                            return ResponseHelper::sendFailedResponse(['message'=>'Bad Request'],400);
+                        }
                     }
-
-                   }
-
+                }
+                if (count($ids) > 0 and $valid_count == 0) {
+                    return ResponseHelper::sendFailedResponse(['message'=>'Bad Request'],400);
+                }
              }else if($question->survey_question_type === SurveyType::TYPE_FILE) {
                  //save multiple
                 if (count($value) > 0 ) {
